relocate test resources
