+++ galen-integration-tests/src/test/java/net/mindengine/galen/tests/integration/ComponentFrameIT.java
@@ -37,7 +37,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class ComponentFrameIntegrationTest {
+public class ComponentFrameIT {
 
 
     private WebDriver driver;
