+++ galen-integration-tests/src/test/resources/sample-test-website/fonts/glyphicons-halflings-regular.eot
