+++ galen-integration-tests/src/test/resources/sample-test-website/js/bootstrap.js
