+++ galen-integration-tests/src/test/resources/sample-test-website/js/handlebars-v1.3.0.js
