+++ galen-integration-tests/src/test/resources/sample-test-website/js/myapp.js
